+++ javalib/src/main/scala/java/lang/Character.scala
@@ -65,11 +65,11 @@ class Character(val _value: scala.Char)
   protected def +(x: String): String = _value + x
 
   protected def <<(x: scala.Int): scala.Int   = _value << x
-  protected def <<(x: scala.Long): scala.Int  = _value << x
+  protected def <<(x: scala.Long): scala.Int  = _value << x.toInt
   protected def >>>(x: scala.Int): scala.Int  = _value >>> x
-  protected def >>>(x: scala.Long): scala.Int = _value >>> x
+  protected def >>>(x: scala.Long): scala.Int = _value >>> x.toInt
   protected def >>(x: scala.Int): scala.Int   = _value >> x
-  protected def >>(x: scala.Long): scala.Int  = _value >> x
+  protected def >>(x: scala.Long): scala.Int  = _value >> x.toInt
 
   protected def ==(x: scala.Byte): scala.Boolean   = _value == x
   protected def ==(x: scala.Short): scala.Boolean  = _value == x
